requirements.txt
numpy
colour
numba
scikit-learn
matplotlib
Pillow
PyYAML